WIP: ts-worker code reorganization
