+++ tradetrust/ts-document-store-worker/src/repos.ts
@@ -1,9 +1,9 @@
-import AWS, {S3, SQS, KMS} from './aws';
+import AWS, {S3, SQS, KMS} from 'src/aws';
 import {
   IBatchedIssueConfig,
   IBatchedDocumentStoreTaskConfig,
   IBatchedRevokeConfig,
-} from './config';
+} from 'src/config';
 
 const S3Service = S3({});
 const SQSService = SQS({});
