+++ tradetrust/ts-document-store-worker/src/tasks/common/errors.ts
