+++ tradetrust/ts-document-store-worker/src/tasks/common/interfaces.ts
