+++ tradetrust/ts-document-store-worker/src/tasks/v2/revoke-batch.ts
@@ -1,9 +1,9 @@
-import { logger } from '../logger';
+import { logger } from 'src/logger';
 import {
   SendDocumentStoreTransaction,
   ISendDocumentStoreTransactionProps
-} from './send-document-store-transaction';
-import { Batch } from './data';
+} from 'src/tasks/common/send-document-store-transaction';
+import { Batch } from 'src/tasks/common/data';
 
 
 interface IRevokeDocumentProps extends ISendDocumentStoreTransactionProps{
