+++ tradetrust/ts-document-store-worker/src/tasks/v2/wrap-batch.ts
@@ -1,12 +1,12 @@
 import { wrapDocuments } from '@govtechsg/open-attestation';
-import { logger } from '../logger';
-import { Batch } from './data';
-import { Task } from './interfaces';
+import { logger } from 'src/logger';
+import { Batch } from 'src/tasks/common/data';
+import { Task } from 'src/tasks/common/interfaces';
+
 
 interface IWrapBatchProps{
   batch: Batch
 }
-
 
 
 class WrapBatch implements Task<void>{
